# docs/02-architecture/api/openapi.yaml
openapi: 3.1.0
info:
  title: Forge Platform API
  version: 1.0.0
  description: >-
    REST contract consumed by the Forge SPA (catalog, approval queue, trace
    viewer). Consistent with the data model (data-model.md) and the agent DNA
    contract (dna-schema.json). Governance is structural: publishing is
    eval-gated (409 when the gate has not passed), approvals expire server-side
    and there is no extend/auto-approve operation, and a role header
    demonstrates segregation of duties (NFR-5). This is a demonstration
    contract, not a production auth design.
servers:
  - url: /api/v1
...
paths:
  /agents:
    get:
      tags: [Agents]
      summary: List agents in the catalog
      parameters:
        - $ref: '#/components/parameters/RoleHeader'
      responses:
        '200':
          description: Agent list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Agent' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Agents]
      summary: Create an agent (identity only; versions are added separately)
      parameters:
        - $ref: '#/components/parameters/RoleHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateAgent' }
      responses:
        '201':
          description: Agent created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Agent' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /agents/{agentId}/versions:
    parameters:
      - $ref: '#/components/parameters/AgentId'
    get:
      tags: [Agents]
      summary: List versions of an agent
      parameters:
        - $ref: '#/components/parameters/RoleHeader'
      responses:
        '200':
          description: Version list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AgentVersion' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [Agents]
      summary: Create a new agent version (draft) from a DNA document
      description: >-
        The dna body is validated against dna-schema.json before the version is
        accepted. Invalid DNA returns 400 with the validation errors in details.
      parameters:
        - $ref: '#/components/parameters/RoleHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateAgentVersion' }
      responses:
        '201':
          description: Draft version created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AgentVersion' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /agents/{agentId}/versions/{version}:
    parameters:
      - $ref: '#/components/parameters/AgentId'
      - $ref: '#/components/parameters/Version'
    get:
      tags: [Agents]
      summary: Get one agent version, including its DNA
      parameters:
        - $ref: '#/components/parameters/RoleHeader'
      responses:
        '200':
          description: Agent version
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AgentVersion' }
        '404': { $ref: '#/components/responses/NotFound' }
  /agents/{agentId}/versions/{version}/publish:
    parameters:
      - $ref: '#/components/parameters/AgentId'
      - $ref: '#/components/parameters/Version'
    post:
      tags: [Agents]
      summary: Publish a version (hard eval gate)
      description: >-
        Transitions draft to published ONLY if the version has a completed eval
        run with passed=true for the suite its own DNA declares
        (evals.suite_ref) — a run of a different version or a different suite
        satisfies nothing (FR-F2). If the gate has not passed, the server
        returns 409 with code publish_gate_unmet and does not publish. The gate
        cannot be bypassed: there is no force flag and no other route to
        status=published. (The seed script's directly-published demo versions
        are the one documented exception, marked by a null
        published_eval_run_id.) Requires the configurator role.
      parameters:
        - $ref: '#/components/parameters/RoleHeader'
      responses:
        '200':
          description: >-
            Version published. published_eval_run_id records the passing run
            that satisfied the gate — the evidence travels with the version.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AgentVersion' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: >-
            Publish gate not satisfied (code publish_gate_unmet) — no completed
            eval run, a failed one, or a declared suite that is not installed;
            the version remains in draft. Also answered (code
            version_not_draft) when the version is already published or
            suspended: only a draft can be published.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example:
                code: publish_gate_unmet
                message: Eval suite has not passed for this version
                details:
                  suite_ref: meridian-ap-eval-suite@1.0.0
                  latest_eval_run_id: null
                  latest_eval_run_passed: false
                  passed_count: 17
                  total: 20
  /agents/{agentId}/versions/{version}/suspend:
    parameters:
      - $ref: '#/components/parameters/AgentId'
      - $ref: '#/components/parameters/Version'
    post:
      tags: [Agents]
      summary: Suspend a published version (manual)
      description: >-
        Halts new runs; history is retained. The transition and its reason are
        recorded as a version.suspended event, and from the next request on a
        run of this version is refused with code agent_suspended (the refusal
        itself recorded as a governance.run_refused event). Needs
        `agent.suspend`, which the configurator and admin roles hold — stopping
        an agent is the fail-safe direction. Resuming is not: see /resume.
      parameters:
        - $ref: '#/components/parameters/RoleHeader'
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SuspendRequest' }
      responses:
        '200':
          description: Version suspended, transition recorded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AgentVersion' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: >-
            Version is not published (code version_not_published) — only a
            published version can be suspended
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /agents/{agentId}/versions/{version}/resume:
    parameters:
      - $ref: '#/components/parameters/AgentId'
      - $ref: '#/components/parameters/Version'
    post:
      tags: [Agents]
      summary: Resume a suspended version (admin only)
      description: >-
        Transitions suspended → published — the one operation that reverses a
        containment, whether the circuit breaker tripped it (FR-G4) or a person
        suspended it by hand. Deliberately manual: there is no cool-down, no
        retry window, and nothing that un-suspends an agent by itself. Needs
        `agent.resume`, which only the admin role holds and which no role
        holding `agent.configure` or `agent.publish` may ever hold — whoever
        built or shipped an agent cannot override the breaker that contained
        it. The resume is recorded as a version.resumed event with its actor,
        optional note, and the suspension it overrode.
      parameters:
        - $ref: '#/components/parameters/RoleHeader'
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResumeRequest' }
      responses:
        '200':
          description: Version resumed, transition recorded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AgentVersion' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: >-
            Version is not suspended (code version_not_suspended) — only a
            suspended version can be resumed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
components:
  parameters:
    RoleHeader:
      name: X-Forge-Role
      in: header
      required: true
      description: >-
        Demonstration role for segregation of duties (NFR-5). configurator
        holds read, agent.configure, agent.publish, agent.suspend and
        run.start; approver holds read and approval.decide; viewer holds read
        only; admin holds read, agent.suspend and agent.resume — the
        containment operator, deliberately not a superuser. No role holds both
        agent.configure/agent.publish and approval.decide, and none holds both
        agent.configure/agent.publish and agent.resume: the person who decides
        what an agent may do is never the person who approves what it proposes,
        and whoever built or shipped an agent can never override the breaker
        that contained it. The build refuses to start if either pairing ever
        stops being true. A role lacking the permission for an operation gets
        403, and the refusal is recorded as a governance.permission_denied
        event. Not authentication.
      schema:
        type: string
        enum: [configurator, approver, viewer, admin]
    AgentId:
      name: agentId
      in: path
      required: true
      schema: { type: string, format: uuid }
    Version:
      name: version
      in: path
      required: true
      description: Semver of the agent version, e.g. 1.0.0
      schema: { type: string }
...
  responses:
    BadRequest:
      description: Malformed request or DNA that fails schema validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: The X-Forge-Role is not permitted to perform this operation (NFR-5)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, description: Stable machine-readable error code }
        message: { type: string, description: Human-readable summary }
        details:
          type: object
          additionalProperties: true
          description: Optional structured context (e.g. validation errors)
    CreateAgent:
      type: object
      required: [slug, name, type]
      properties:
        slug: { type: string, pattern: '^[a-z0-9]+(-[a-z0-9]+)*$' }
        name: { type: string }
        type: { type: string, enum: [chatbot, workflow, autonomous] }
        description: { type: string }
    Agent:
      type: object
      required: [id, tenant_id, slug, name, type]
      properties:
        id: { type: string, format: uuid }
        tenant_id: { type: string }
        slug: { type: string }
        name: { type: string }
        type: { type: string, enum: [chatbot, workflow, autonomous] }
        description: { type: string }
        created_at: { type: string, format: date-time }
    CreateAgentVersion:
      type: object
      required: [dna]
      properties:
        dna:
          $ref: '#/components/schemas/Dna'
    Dna:
      type: object
      description: >-
        A complete agent DNA document. The authoritative structure is
        dna-schema.json (JSON Schema draft 2020-12); it is validated there at
        write time and is not re-declared field by field here so the contract
        stays single-sourced. Top-level blocks: identity, instructions, tools,
        knowledge, model, guardrails, evals. Fail-closed governance fields
        (escalate_on_no_rule_match, require_citations, publish_gate,
        authority_policy) are const-locked and cannot be disabled.
      additionalProperties: true
      externalDocs:
        description: Agent DNA JSON Schema
        url: ../dna-schema.json
    AgentVersion:
      type: object
      required: [id, tenant_id, agent_id, version, status, dna]
      properties:
        id: { type: string, format: uuid }
        tenant_id: { type: string }
        agent_id: { type: string, format: uuid }
        version: { type: string, description: semver }
        status: { type: string, enum: [draft, published, suspended] }
        dna: { $ref: '#/components/schemas/Dna' }
        published_eval_run_id:
          type: [string, 'null']
          format: uuid
          description: The passing eval run that satisfied the publish gate, if published
        published_at: { type: [string, 'null'], format: date-time }
        created_at: { type: string, format: date-time }
    SuspendRequest:
      type: object
      properties:
        reason:
          type: string
          description: Why, recorded verbatim in the version.suspended event
    ResumeRequest:
      type: object
      properties:
        note:
          type: string
          description: Recorded verbatim in the version.resumed event, with the actor
...